requirements.txt
Flask==2.3.2
PyMuPDF==1.22.3
python-docx==1.1.0
scikit-learn==1.2.2
joblib==1.3.2
